clean up hover tests
