+++ packages/language-server/src/__test__/__fixtures__/multi-file/user-posts/User.prisma
@@ -4,4 +4,6 @@ model User {
     name  String
     email String
     posts Post[]
+
+    favouriteAnimal FavouriteAnimal
 }
